Optimized adb connection for app launch at AVM boot time.
It's observed that app launch might be not working when user click
app right after the OS boot up. The root cause is the CIV might be
running into paused status or adb is killed by other process for
power saving. The fixes are:
1) Don't pause CIV and kill adb when Android app is running.
2) Establish the adb & resume civ if user intends to launch android
   app.
3) Establish the adb & resume civ if user intends to launch android
   app.

Added TimeoutStopSec to civ.service to ensure systemd could send
sigkill to qemu process if sigterm is not able to kill the qemu
process.

Signed-off-by: Wan Shuang <shuang.wan@intel.com>